Added option for fixed viewing azimuth angle
Added an option (dynamic_azimuth) to compute the variable "psi"
(the azimuthal difference between solar and observation angle)
internally in time-series mode based on a fixed observation viewing
angle (called "vazi" in the input data). This modification calls a
new function called calcazimuthangle.m that computes the solar
azimuth angle and psi, the relative difference between that and the
observation angle.